+++ gulpfile.js
@@ -3,9 +3,6 @@ const sass = require('gulp-sass');
 const ts = require('gulp-typescript');
 const del = require('delete');
 const cleanCss = require('gulp-clean-css');
-const renderer = require('./dist/Renderer');
-const fsx = require("fs-extra");
-const path = require("path");
 
 function clearDist(cb) {
     del('dist/*', cb);
@@ -25,18 +22,10 @@ function compileSass() {
         .pipe(dest('dist/styles'));
 }
 
-async function renderMarkdown() {
-    let filename = 'test.md';
-    let rend = new renderer.Renderer();
-    let html = await rend.render(filename);
-    await fsx.writeFile(filename.replace(path.extname(filename), '') + '.html', html);
-}
-
 task('cleanBuild', series(clearDist, compileTypescript, compileSass));
 task('default', series(compileTypescript, compileSass));
 task('watch', () => {
-    series(compileSass, compileTypescript, renderMarkdown);
-    watch('src/**/*.sass', series(compileSass, renderMarkdown));
+    series(compileSass, compileTypescript);
+    watch('src/**/*.sass', series(compileSass));
     watch('**/*.ts', compileTypescript);
-    watch('**/*.md', renderMarkdown);
 });