+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 - Changelog
 - markdown parsing with markdown it
 - command to include plugins `[!use]: plugin`
-- command to include markdown files `[!include]: file`
+- command to include markdown files `[!include]: file.md`
 - command to start a new page `[!newpage]`
+- command to include a stylesheet `[!stylesheet]: file.css`
 - option to export to pdf `--pdf`
 - option to watch (and export to html) `--watch`
 - stylesheets supporting math and code highlighting (with highlightjs)
+- auto base64 converting for images for a standalone html